Pull subscriber method test env. to sub-package level
