+++ server/src/test/java/io/spine/server/event/model/EventSubscriberMethodTest.java
@@ -23,17 +23,17 @@ package io.spine.server.event.model;
 import com.google.common.testing.NullPointerTester;
 import com.google.protobuf.Any;
 import io.spine.core.EventContext;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.ARejectionSubscriber;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.InvalidNoAnnotation;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.InvalidNoParams;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.InvalidNotVoid;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.InvalidOneNotMsgParam;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.InvalidTooManyParams;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.InvalidTwoParamsFirstInvalid;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.InvalidTwoParamsSecondInvalid;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.ValidButPrivate;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.ValidOneParam;
-import io.spine.server.event.given.EventSubscriberMethodTestEnv.ValidTwoParams;
+import io.spine.server.event.model.given.subscriber.ARejectionSubscriber;
+import io.spine.server.event.model.given.subscriber.InvalidNoAnnotation;
+import io.spine.server.event.model.given.subscriber.InvalidNoParams;
+import io.spine.server.event.model.given.subscriber.InvalidNotVoid;
+import io.spine.server.event.model.given.subscriber.InvalidOneNotMsgParam;
+import io.spine.server.event.model.given.subscriber.InvalidTooManyParams;
+import io.spine.server.event.model.given.subscriber.InvalidTwoParamsFirstInvalid;
+import io.spine.server.event.model.given.subscriber.InvalidTwoParamsSecondInvalid;
+import io.spine.server.event.model.given.subscriber.ValidButPrivate;
+import io.spine.server.event.model.given.subscriber.ValidOneParam;
+import io.spine.server.event.model.given.subscriber.ValidTwoParams;
 import io.spine.server.model.given.Given;
 import io.spine.test.reflect.event.RefProjectCreated;
 import org.junit.jupiter.api.DisplayName;
